Assignment 2.1 Cable Cost Calculator
Using comments, create a header at the top of the program indicating the purpose of the program, assignment number, and your name. Use the SIUE Style Guide as a reference.
Display a welcome message for your user.
Retrieve the company name from the user.
Retrieve the number of feet of fiber optic cable to be installed from the user.
Calculate the installation cost of fiber optic cable by multiplying the total cost as the number of feet times $0.87.
Print a receipt for the user including the company name, number of feet of fiber to be installed, the calculated cost, and total cost in a legible format.
Include appropriate comments throughout the program.